+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 :pencil: - chore
 :microscope: - experimental
 
+## [0.46.3]
+- :rocket: added _I refresh page until text of {string} {playwrightValidation} {string}( ){playwrightTimeout}_ step
+- :rocket: added _I refresh page until {string} {playwrightConditionWait}( ){playwrightTimeout}_ step
+
 ## [0.46.2]
 - :beetle: added error handling during trace attach
 - :pencil: updated dependencies